refactor: split MCP server construction
Move transport-neutral MCP server and instruction wiring under src/mcp while keeping the CLI command focused on stdio startup and setup output.
